+++ game_patch/misc/alpine_settings.cpp
@@ -1220,7 +1220,7 @@ bool alpine_player_settings_load(rf::Player* player)
         processed_keys.insert("GamepadGyroSpace");
     }
     if (settings.count("GamepadGyroModifierMode")) {
-        g_alpine_game_config.gamepad_gyro_modifier_mode = std::clamp(std::stoi(settings["GamepadGyroModifierMode"]), 0, 5);
+        g_alpine_game_config.gamepad_gyro_ratcheting = std::clamp(std::stoi(settings["GamepadGyroModifierMode"]), 0, 5);
         processed_keys.insert("GamepadGyroModifierMode");
     }
     if (settings.count("GamepadGyroInvertX")) {
@@ -1531,7 +1531,7 @@ void alpine_control_config_serialize(std::ofstream& file, const rf::ControlConfi
     file << "GamepadGyroScannerSensitivityModifier=" << g_alpine_game_config.gamepad_scanner_gyro_sensitivity_modifier << "\n";
     file << "GamepadGyroAutocalibrationMode=" << g_alpine_game_config.gamepad_gyro_autocalibration_mode << "\n";
     file << "GamepadGyroSpace=" << g_alpine_game_config.gamepad_gyro_space << "\n";
-    file << "GamepadGyroModifierMode=" << g_alpine_game_config.gamepad_gyro_modifier_mode << "\n";
+    file << "GamepadGyroRatchetingMode=" << g_alpine_game_config.gamepad_gyro_ratcheting << "\n";
     file << "GamepadGyroInvertX=" << g_alpine_game_config.gamepad_gyro_invert_x << "\n";
     file << "GamepadGyroInvertY=" << g_alpine_game_config.gamepad_gyro_invert_y << "\n";
     file << "GamepadGyroTightening=" << g_alpine_game_config.gamepad_gyro_tightening << "\n";