+++ game_patch/misc/ui.cpp
@@ -289,10 +289,10 @@ static rf::ui::Checkbox ao_gyro_autocalibration_cbox;
 static rf::ui::Label ao_gyro_autocalibration_label;
 static rf::ui::Label ao_gyro_autocalibration_butlabel;
 static char ao_gyro_autocalibration_butlabel_text[16];
-static rf::ui::Checkbox ao_gyro_modifier_mode_cbox;
-static rf::ui::Label ao_gyro_modifier_mode_label;
-static rf::ui::Label ao_gyro_modifier_mode_butlabel;
-static char ao_gyro_modifier_mode_butlabel_text[12];
+static rf::ui::Checkbox ao_gyro_ratcheting_cbox;
+static rf::ui::Label ao_gyro_ratcheting_label;
+static rf::ui::Label ao_gyro_ratcheting_butlabel;
+static char ao_gyro_ratcheting_butlabel_text[12];
 static rf::ui::Checkbox ao_gyro_space_cbox;
 static rf::ui::Label ao_gyro_space_label;
 static rf::ui::Label ao_gyro_space_butlabel;
@@ -1124,8 +1124,8 @@ void ao_gyro_autocalibration_cbox_on_click([[maybe_unused]] int x, [[maybe_unuse
     ao_play_button_snd(true);
 }
 
-void ao_gyro_modifier_mode_cbox_on_click([[maybe_unused]] int x, [[maybe_unused]] int y) {
-    g_alpine_game_config.gamepad_gyro_modifier_mode = (g_alpine_game_config.gamepad_gyro_modifier_mode + 1) % 6;
+void ao_gyro_ratcheting_cbox_on_click([[maybe_unused]] int x, [[maybe_unused]] int y) {
+    g_alpine_game_config.gamepad_gyro_ratcheting = (g_alpine_game_config.gamepad_gyro_ratcheting + 1) % 6;
     ao_play_button_snd(true);
 }
 
@@ -1992,9 +1992,9 @@ void alpine_options_panel_init() {
     alpine_options_panel_inputbox_init(
         &ao_joy_scopesens_cbox, &ao_joy_scopesens_label, &ao_joy_scopesens_butlabel, &alpine_options_panel2, ao_joy_scopesens_cbox_on_click, 112, 354, "Joy scope mod");
     alpine_options_panel_checkbox_init(
-        &ao_gyro_enabled_cbox, &ao_gyro_enabled_label, &alpine_options_panel2, ao_gyro_enabled_cbox_on_click, g_alpine_game_config.gamepad_gyro_enabled, 280, 204, "Gyro aiming");
+        &ao_gyro_enabled_cbox, &ao_gyro_enabled_label, &alpine_options_panel2, ao_gyro_enabled_cbox_on_click, g_alpine_game_config.gamepad_gyro_enabled, 280, 204, "Gyro camera");
     alpine_options_panel_inputbox_init(
-        &ao_gyro_sensitivity_cbox, &ao_gyro_sensitivity_label, &ao_gyro_sensitivity_butlabel, &alpine_options_panel2, ao_gyro_sensitivity_cbox_on_click, 280, 234, "Gyro sensitivity");
+        &ao_gyro_sensitivity_cbox, &ao_gyro_sensitivity_label, &ao_gyro_sensitivity_butlabel, &alpine_options_panel2, ao_gyro_sensitivity_cbox_on_click, 280, 234, "Gyro cam sensitivity");
     alpine_options_panel_checkbox_init(
         &ao_gyro_invert_x_cbox, &ao_gyro_invert_x_label,
         &alpine_options_panel2, ao_gyro_invert_x_cbox_on_click,
@@ -2014,8 +2014,8 @@ void alpine_options_panel_init() {
         &ao_gyro_autocalibration_cbox, &ao_gyro_autocalibration_label, &ao_gyro_autocalibration_butlabel,
         &alpine_options_panel2, ao_gyro_autocalibration_cbox_on_click, 280, 324, "Gyro auto-calib");
     alpine_options_panel_inputbox_init(
-        &ao_gyro_modifier_mode_cbox, &ao_gyro_modifier_mode_label, &ao_gyro_modifier_mode_butlabel,
-        &alpine_options_panel2, ao_gyro_modifier_mode_cbox_on_click, 280, 354, "Gyro modifier");
+        &ao_gyro_ratcheting_cbox, &ao_gyro_ratcheting_label, &ao_gyro_ratcheting_butlabel,
+        &alpine_options_panel2, ao_gyro_ratcheting_cbox_on_click, 280, 354, "Gyro ratcheting");
     alpine_options_panel_inputbox_init(
         &ao_gyro_tightening_cbox, &ao_gyro_tightening_label, &ao_gyro_tightening_butlabel,
         &alpine_options_panel2, ao_gyro_tightening_cbox_on_click, 280, 354, "Gyro tightening");
@@ -2319,11 +2319,11 @@ void alpine_options_panel_do_frame(int x)
     ao_gyro_autocalibration_butlabel.text  = ao_gyro_autocalibration_butlabel_text;
     ao_gyro_autocalibration_butlabel.align = rf::gr::ALIGN_CENTER;
 
-    static const char* gyro_modifier_mode_names[] = {"Always", "Hold (Off)", "Hold (On)", "Toggle", "Touch (On)", "Touch (Off)"};
-    snprintf(ao_gyro_modifier_mode_butlabel_text, sizeof(ao_gyro_modifier_mode_butlabel_text), "%s",
-        gyro_modifier_mode_names[std::clamp(g_alpine_game_config.gamepad_gyro_modifier_mode, 0, 5)]);
-    ao_gyro_modifier_mode_butlabel.text  = ao_gyro_modifier_mode_butlabel_text;
-    ao_gyro_modifier_mode_butlabel.align = rf::gr::ALIGN_CENTER;
+    static const char* gyro_ratcheting_names[] = {"Always", "Hold (Off)", "Hold (On)", "Toggle", "Touch (On)", "Touch (Off)"};
+    snprintf(ao_gyro_ratcheting_butlabel_text, sizeof(ao_gyro_ratcheting_butlabel_text), "%s",
+        gyro_ratcheting_names[std::clamp(g_alpine_game_config.gamepad_gyro_ratcheting, 0, 5)]);
+    ao_gyro_ratcheting_butlabel.text  = ao_gyro_ratcheting_butlabel_text;
+    ao_gyro_ratcheting_butlabel.align = rf::gr::ALIGN_CENTER;
 
     snprintf(ao_gyro_space_butlabel_text, sizeof(ao_gyro_space_butlabel_text), "%s", gyro_get_space_name(g_alpine_game_config.gamepad_gyro_space));
     ao_gyro_space_butlabel.text  = ao_gyro_space_butlabel_text;
@@ -2365,7 +2365,7 @@ void alpine_options_panel_do_frame(int x)
     ao_gyro_vh_mixer_butlabel.x          = 280 + 50;
     ao_gyro_space_butlabel.x             = 280 + 50;
     ao_gyro_autocalibration_butlabel.x   = 280 + 50;
-    ao_gyro_modifier_mode_butlabel.x     = 280 + 50;
+    ao_gyro_ratcheting_butlabel.x     = 280 + 50;
     ao_rumble_filter_butlabel.x          = 280 + 50;
     ao_gamepad_icon_override_butlabel.x  = 280 + 50;
     ao_input_prompt_mode_butlabel.x      = 280 + 50;
@@ -2385,9 +2385,9 @@ void alpine_options_panel_do_frame(int x)
     ao_gyro_autocalibration_label.enabled  = gyro_hw;
     ao_gyro_autocalibration_butlabel.enabled = gyro_hw;
 
-    ao_gyro_modifier_mode_cbox.enabled     = gyro_enabled;
-    ao_gyro_modifier_mode_label.enabled    = gyro_enabled;
-    ao_gyro_modifier_mode_butlabel.enabled = gyro_enabled;
+    ao_gyro_ratcheting_cbox.enabled     = gyro_enabled;
+    ao_gyro_ratcheting_label.enabled    = gyro_enabled;
+    ao_gyro_ratcheting_butlabel.enabled = gyro_enabled;
 
     ao_gyro_invert_x_cbox.enabled        = gyro_enabled;
     ao_gyro_invert_x_label.enabled       = gyro_enabled;
@@ -2446,7 +2446,7 @@ void alpine_options_panel_do_frame(int x)
         rc.add_inputbox(ao_gyro_sensitivity_cbox, ao_gyro_sensitivity_label, ao_gyro_sensitivity_butlabel, gyro_enabled);
         rc.add_inputbox(ao_gyro_scopesens_cbox, ao_gyro_scopesens_label, ao_gyro_scopesens_butlabel, gyro_enabled);
         rc.add_inputbox(ao_gyro_scannersens_cbox, ao_gyro_scannersens_label, ao_gyro_scannersens_butlabel, gyro_enabled);
-        rc.add_inputbox(ao_gyro_modifier_mode_cbox, ao_gyro_modifier_mode_label, ao_gyro_modifier_mode_butlabel, gyro_enabled);
+        rc.add_inputbox(ao_gyro_ratcheting_cbox, ao_gyro_ratcheting_label, ao_gyro_ratcheting_butlabel, gyro_enabled);
         rc.add_inputbox(ao_gyro_tightening_cbox, ao_gyro_tightening_label, ao_gyro_tightening_butlabel, gyro_enabled);
         rc.add_inputbox(ao_gyro_smoothing_cbox, ao_gyro_smoothing_label, ao_gyro_smoothing_butlabel, gyro_enabled);
         rc.add_inputbox(ao_gyro_vh_mixer_cbox, ao_gyro_vh_mixer_label, ao_gyro_vh_mixer_butlabel, gyro_enabled);